gwhttpd: Explicitly set sa_family before bind()
Yonle reported that he is getting an error on bind():

  Address family not supported by protocol

This doesn't mean that Yonle's device doesn't support IPv4, but it's
because we don't explicitly set the sa_family to AF_INET. This will
happen to work just fine only on a platform where `AF_INET == 0` since
we have a prior memset() call that zeroes the whole struct sockaddr_in.

On platforms that have `AF_INET != 0`, this will break. Fix this by
simply setting the `saddr.sa_family` to AF_INET at initialization.

Fixes: https://github.com/ammarfaizi2/gwhttpd/issues/1
Link: https://t.me/GNUWeeb/627604
Reported-by: Yonle <yonle@gnuweeb.org>
Tested-by: Yonle <yonle@gnuweeb.org>
Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>