gwhttpd: Fix `EBADF` errors at destruction
When Yonle reported an issue about error on bind(), he attached an
strace output and it has many EBADF errors like this:

  close(0) = 0
  close(0) = -1 EBADF (Bad file descriptor)
  close(0) = -1 EBADF (Bad file descriptor)
  close(0) = -1 EBADF (Bad file descriptor)
  close(0) = -1 EBADF (Bad file descriptor)
  close(0) = -1 EBADF (Bad file descriptor)
  ...

This happens because all sess[i].fd is set to 0 at initialization while
the destructor is supposed to close the file descriptor when it's not
equal to -1. Fix this by setting all file descriptor fields to -1 at
initialization.

Link: https://github.com/ammarfaizi2/gwhttpd/issues/1
Reported-by: Yonle <yonle@gnuweeb.org>
Tested-by: Yonle <yonle@gnuweeb.org>
Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>