gwhttpd: Avoid endless busy spinning on `send()`
When we fail to send() due to EAGAIN, we retry the send(). However,
if we are spinning on this retry for so long, this will eat CPU cycle
and slow down the entire application. We should stop retrying at
some point. Add a loop counter and return -ENETDOWN if we are failing
too many times in the send() retry loop cycle.

Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>