gwhttpd: Fix 403 HTTP error when accessing an empty file
Hendra reports that he is seeing an HTTP 403 error when accssing an
empty file. It turned out that we fail on the mmap() syscall. mmap()
returns -EINVAL if the given size is zero. This happens when the file
size is zero. Make a special case when the file size is zero, bump the
mmap() size to 1. This way the mmap() will happily allocate a VMA
without error.

Link: https://t.me/GNUWeeb/634242
Tested-by: Hendra Manudinata <saya@hendra-manudinata.my.id>
Reported-by: Hendra Manudinata <saya@hendra-manudinata.my.id>
Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>